ProcessInfo

+++ TapBox/OnScreenKeyboard.cs
@@ -4,23 +4,66 @@ namespace TapBox
 
     public static class OnScreenKeyboard
     {
-        private const string TouchKeyboardPath = @"C:\Program Files\Common Files\Microsoft Shared\Ink\TabTip.exe";
-        private static readonly string ProcessName = System.IO.Path.GetFileNameWithoutExtension(TouchKeyboardPath);
-        private static readonly ProcessStartInfo ProcessStartInfo = new ProcessStartInfo(TouchKeyboardPath);
+        private static ProcessInfo _processInfo = ProcessInfo.Create(@"C:\Program Files\Common Files\Microsoft Shared\Ink\TabTip.exe");
+
+        public static string TouchKeyboardPath
+        {
+            get { return _processInfo?.ProcessStartInfo.FileName; }
+            set { _processInfo = ProcessInfo.Create(value); }
+        }
 
         public static void Show()
         {
+            if (_processInfo == null)
+            {
+                return;
+            }
+
             Debug.WriteLine("show");
-            Process.Start(ProcessStartInfo);
+            Process.Start(_processInfo.ProcessStartInfo);
         }
 
         public static void Hide()
         {
+            if (_processInfo == null)
+            {
+                return;
+            }
+
             Debug.WriteLine("hide");
-            foreach (var process in Process.GetProcessesByName(ProcessName))
+            foreach (var process in Process.GetProcessesByName(_processInfo.ProcessName))
             {
                 process.Kill();
             }
+        }
+
+        private class ProcessInfo
+        {
+            private ProcessInfo(string path)
+            {
+                ProcessStartInfo = new ProcessStartInfo(path);
+                ProcessName = System.IO.Path.GetFileNameWithoutExtension(path);
+            }
+
+            public ProcessStartInfo ProcessStartInfo { get; }
+
+            public string ProcessName { get; }
+
+            public static ProcessInfo Create(string path)
+            {
+                if (path == null)
+                {
+                    return null;
+                }
+
+                if (System.IO.File.Exists(path))
+                {
+                    return new ProcessInfo(path);
+                }
+
+                return null;
+            }
+
         }
     }
 }
